Server: backport expired message cleaner updates
Widens the period of the cleaner to be 12h when row counts cleaned dips
below the batch size.

The legacy cleaner query also has its timeout extended as was done when
troubleshooting the cleaner for one deployment. We may need to do the
same for the new `messagecontent` table, but for now I left it as-is.

Some additional refactors were needed with the 12h period to ensure the
process would not block (for hours) when the shutdown atomic gets set.

Additionally, since the "worst case" search for the `message` table
cleaner is the inevitable conclusion when a deployment starts writing to
`messagecontent` instead, a check is done when the cleaner first starts
up to see if there are any pending payloads at all in the `message` table.

This allows us to skip repeatedly scanning a table that will never have new
message payloads added to it.